update balances balance colunm type

+++ database/migrations/2021_02_25_201220_create_balances_table.php
@@ -15,7 +15,7 @@ class CreateBalancesTable extends Migration
     {
         Schema::create('balances', function (Blueprint $table) {
             $table->id();
-            $table->integer('balance');
+            $table->bigInteger('balance');
             $table->foreignId('user_id')->constrained('users');
             $table->timestamps();
         });
